move materials to completed
